Quick-wins batch: local-chat BUG filter + 6 code-sweep fixes
FEEDBACK-3: local-chat dev-feedback log + auto-ack now only fire on lines
carrying the uppercase token "BUG", so ordinary chatter no longer fills
feedback.log or trips the auto-ack. Validated end to end (feedback_probe.py):
plain line ignored, BUG line logged+acked, second BUG line throttled.

Code-sweep fixes (found by a scan, each 1-few lines):
- LOOP-1 (eve-server.cpp): main loop slept the elapsed frame time instead of
  the remaining budget (m_sleepTime - elapsed) -> fast frames spun hot.
- STATION-1 (Station.cpp): GetOfficeID used `=` not `==`, returning the first
  office for any corp.
- PLAYER-1 (SystemManager.cpp): dead `if (m_players < 0)` clamp on a uint16
  (underflow wrapped to 65535); now checks ==0 before decrement.
- MISSION-DEST-1 (MapData.cpp): GetMissionDestination could underflow
  sysList.size()-1 and throw from .at() (node crash); added empty-list guard.
- FMT-1/2 (AlertService.cpp, LSCService.cpp): %u fed a size_t; cast to uint32.

Deployed and verified stable (no player-count desync, loop CPU idle ~1%).

+++ src/eve-server/admin/AlertService.cpp
@@ -87,7 +87,7 @@ PyResult AlertService::BeanDelivery(PyCallArgs& call, PyList* beans)
 
 PyResult AlertService::GroupBeanDelivery(PyCallArgs& call, PyBuffer* compressedBeans)
 {
-    _log(CLIENT__WARNING, "AlertService::Handle_GroupBeanDelivery(): size=%u", call.tuple->size() );
+    _log(CLIENT__WARNING, "AlertService::Handle_GroupBeanDelivery(): size=%u", (uint32)call.tuple->size() );
     //call.Dump(CLIENT__CALL_DUMP);
 
     return PyStatic.NewNone();
