+++ src/eve-server/chat/LSCService.cpp
@@ -376,7 +376,7 @@ PyResult LSCService::SendMessage(PyCallArgs& call, PyRep* channelInfo, PyWString
         PyTuple* tuple0 = call.tuple->AsTuple();
 
         if (tuple0->size() != 2) {
-            _log(LSC__ERROR, "LSCService::Handle_SendMessage failed: tuple0 is the wrong size: expected 2, but got %u", tuple0->size());
+            _log(LSC__ERROR, "LSCService::Handle_SendMessage failed: tuple0 is the wrong size: expected 2, but got %u", (uint32)tuple0->size());
             return PyStatic.NewNone();
         }
